api setup

+++ migrations/001_initial_schema.up.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS users (
 
 CREATE TABLE IF NOT EXISTS accounts (
   id BIGSERIAL PRIMARY KEY,
-  user_id INT REFERENCES users(id) ON DELETE CASCADE,
+  user_id BIGINT REFERENCES users(id) ON DELETE CASCADE,
   balance BIGINT DEFAULT 0,
   created_at BIGINT DEFAULT (floor(extract(epoch FROM now() AT TIME ZONE 'UTC') * 1000)::BIGINT)
 );
@@ -18,7 +18,7 @@ INSERT INTO transaction_types (type) VALUES ('deposit'), ('withdrawal');
 
 CREATE TABLE IF NOT EXISTS transactions (
   id BIGSERIAL PRIMARY KEY,
-  account_id INT REFERENCES accounts(id) ON DELETE CASCADE,
+  account_id BIGINT REFERENCES accounts(id) ON DELETE CASCADE,
   amount BIGINT NOT NULL,
   transaction_type SMALLINT NOT NULL REFERENCES transaction_types(id),
   created_at BIGINT DEFAULT (floor(extract(epoch FROM now() AT TIME ZONE 'UTC') * 1000)::BIGINT)
